requirements.txt
google-generativeai
Pillow
geocoder
feedparser
requests
streamlit-lottie
